# Improve performance of `ren_draw_rect`