We color a 1x1 `SDL_Surface` with the desired color. This surface is then stretched over the area we need to cover using `SDL_BlitScaled`.

This way we avoid having to do the blending ourselves.