@jgmdev 2022-04-19
Tested this change and everything is working normally, would be nice to have some figures on the performance gains.