@Guldoman 2022-04-20
Ok, so using this as a benchmark:
```lua
local tt
local tot = 0
local n = 0

local ef = renderer.end_frame
function renderer.end_frame()
  local t = system.get_time()
  if not tt then tt = t end
  ef()
  local e = system.get_time()
  tot = tot + e - t
  n = (n + 1) % 100
  if n == 0 then
    print(string.format("%.4f\t%.4f", tot/100, e - tt))
    tt = e
    tot = 0
  end
end

local clr = { 0, 0, 0, 20 }

local rvd = core.root_view.draw
function core.root_view:draw()
  rvd(self)
  renderer.draw_rect(0, 0, core.root_view.size.x, core.root_view.size.y, clr)
  clr[1] = (clr[1] + 1) % 255
  core.redraw = true
end
```
which basically draws a transparent rectangle over the whole screen and measures how much it takes to draw 100 frames.

**Note**: by drawing the rectangle changing color each frame, we force the redraw of the whole window. This means that the time includes also drawing the interface.
This shouldn't affect the benchmark times because the interface doesn't use transparent rectangles but only completely opaque ones, which aren't really touched by this PR.
So the time it takes to draw the interface is the same time in both cases, and the difference is due to the drawing of the transparent rectangle.

The results were:
Before:
| Average (ms) | 100 frames (ms) |
|--------------|-----------------|
| 0.0242       | 2.4900          |
| 0.0248       | 2.5487          |
| 0.0264       | 2.7106          |
| 0.0235       | 2.4078          |
| 0.0245       | 2.5085          |
| 0.0256       | 2.6342          |
| 0.0246       | 2.5337          |
| 0.0244       | 2.5031          |
| 0.0242       | 2.4848          |
| 0.0241       | 2.4782          |

After:
| Average (ms) | 100 frames (ms) |
|--------------|-----------------|
| 0.0163       | 1.7272          |
| 0.0158       | 1.6764          |
| 0.0156       | 1.6755          |
| 0.0157       | 1.6638          |
| 0.0158       | 1.6788          |
| 0.0157       | 1.6693          |
| 0.0154       | 1.6592          |
| 0.0160       | 1.7253          |
| 0.0158       | 1.6965          |
| 0.0158       | 1.6818          |